working on modualizing my code
