+++ src/javascripts/buttonEvents.js
@@ -1,0 +1,3 @@
+const buttonEvents = () => {
+  document.querySelector('#submitBtn').addEventListener('click', formContent);
+};
